+++ tests/integration/test_load_and_run_agent.py
@@ -7,9 +7,7 @@ from pathlib import Path
 from typing import Any
 
 import pytest
-from litellm import BaseModel
 from litellm.utils import validate_environment
-from pydantic import ConfigDict
 
 from any_agent import AgentConfig, AgentFramework, AnyAgent, TracingConfig
 from any_agent.config import MCPStdio
@@ -225,22 +223,3 @@ def test_load_and_run_agent(
         assert_eval(agent_trace)
     finally:
         agent.exit()
-
-
-def test_output_type(agent_framework: AgentFramework) -> None:
-    class TestOutput(BaseModel):
-        model_config = ConfigDict(extra="forbid")
-        city_name: str
-
-    agent = AnyAgent.create(
-        agent_framework,
-        AgentConfig(
-            model_id="gpt-4.1-mini",
-            output_type=TestOutput,
-        ),
-        tracing=TracingConfig(console=True),
-    )
-
-    result = agent.run("What is the capital of France?")
-    assert isinstance(result.final_output, TestOutput)
-    assert result.final_output.city_name == "Paris"
